requirements-dev.txt
-rrequirements.txt
Sphinx
alabaster
cov-core
coverage
flake8
flake8-bugbear
freezegun
py
pytest
pytest-cov
pytest-mock
pytest-xdist
pytest-flask
pytest-playwright
playwright
psutil
responses
tox>=3.0.0
